clinker-core: split executor tests that assert on private symbols, keep those inline
Three executor test files mix end-to-end integration cases with white-box
unit assertions that read crate-private symbols the tests/ integration
crate cannot see. Splitting them lets the integration cases relocate while
the private-symbol assertions stay where they can reach the symbols they
test, without widening any visibility to move a test.

- aggregation.rs: the two full-pipeline dispatch tests relocate to
  tests/aggregation_dispatch.rs; the hash/streaming accumulator internals,
  the single_predecessor malformed-DAG check, the two-phase byte encoder,
  and the group-boundary sort-order tests stay inline — they reach
  pub(crate) symbols (single_predecessor, group_by_sort_fields,
  GroupBoundary, StreamingErrorMode).

- multi_output.rs: the multi-writer / DLQ-staging integration tests
  relocate to tests/multi_output.rs; the route-evaluation unit tests stay
  inline because they compile and evaluate routes through the private
  PipelineExecutor::compile_route and pub(crate) CompiledRoute::evaluate.

- deferred_dispatch.rs: the commit-pass emission, deferred-buffer E310,
  build-side replay, composition-body harvest, and output-fanout tests
  relocate to tests/deferred_dispatch.rs through the public in-context
  entry point; the deferred-region plan-structure tests
  (deferred_region_at_producer / BoundBody::deferred_regions), the
  RSS-overshoot abort (the pre-seeded MemoryArbitrator run entry), and the
  cascading-retraction loop cap (commit::with_test_loop_cap) stay inline.

scheduling.rs stays entirely inline — every test reads scheduled_pass_order
or the pre-seeded-arbitrator run entry. The executor mod-test module sheds
the relocated submodules and its run_test helper (now run_config in
tests/common); the surviving inline submodules are exactly the white-box
ones.

No assertion changed; every relocated test runs with the same body, and
every white-box symbol keeps its original visibility.

Closes #263, #132
